requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
Faker==35.2.0
inflection==0.5.1
Markdown==3.7
packaging==24.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
validate-docbr==1.10.0
